+++ app/scripts/drsApp.config.js
@@ -6,13 +6,15 @@ import es from '../i18n/drs_es.json';
 import ja from '../i18n/drs_ja.json';
 import ptBr from '../i18n/drs_pt_BR.json';
 
-config.$inject = ['$compileProvider', '$translateProvider', '$translateMessageFormatInterpolationProvider', 'localStorageServiceProvider'];
+config.$inject = ['$compileProvider', '$locationProvider', '$translateProvider', '$translateMessageFormatInterpolationProvider', 'localStorageServiceProvider'];
 
-function config($compileProvider, $translateProvider, $translateMessageFormatInterpolationProvider, localStorageServiceProvider) {
+function config($compileProvider, $locationProvider, $translateProvider, $translateMessageFormatInterpolationProvider, localStorageServiceProvider) {
   // TODO: remove this depenency by fixing component bindings https://github.com/angular/angular.js/blob/master/CHANGELOG.md#breaking-changes-1
   $compileProvider.preAssignBindingsEnabled(true);
   // Allow chrome-extension: URLs in ng-src
   $compileProvider.imgSrcSanitizationWhitelist(/^\s*((https?|chrome-extension):|data:image\/)/);
+
+  $locationProvider.hashPrefix('');
 
   // See https://angular-translate.github.io/docs/#/guide
   $translateProvider.useSanitizeValueStrategy('escape');